move uvdpa client from sriov-network-device-plugin
Inverse the dependency. Instead of depending on
sriov-network-device-plugin, make the client library a subpackage and
base it on the internal json-rpc implementation.

This commit, brings the new uvdpa subpackage:
uvdpa/userClient.go: contains the interface of uvdpa implementations
uvdpa/mockClient.go: contains a mock implementation useful for testing
uvdpa/dpdkVdpaC.go: contains a json-rpc v2.0 implementation that
connects to the dpdk userspace daemon as it is in [1]

[1] https://gitlab.com/mcoquelin/userspace-vdpa

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>